Make Distribution.Extra.Doctest -Wall-clean
An unused `autogenModulesDir` import causes `doctest` to fail to
build with `Cabal-3.0.0.0`, so I decided that it was time to just
bite the bullet and make the code in `doctest` compile without
warnings in `-Wall`. Thankfully, accomplishing this feat proved not
to difficult. There's one mysterious pattern-match coverage check
warning related to `parseComponentName` that _only_ occurs on GHC
7.0, but I strongly suspect that to be due to an old GHC bug, so I
decided to simply turn on `-fno-warn-overlapping-patterns` when
building with 7.0.