Filter out Build_doctests (can't have doctests)
We used to only get module sources from library and executable components.  So
the `Build_doctests` module wouldn't show up.  But for issue #57 we're now
getting module sources for the test suites as well, and Cabal nags us with the
warning:

    <no location info>: warning: [-Wmissing-home-modules]
	These modules are needed for compilation but not listed in your .cabal
        file's other-modules:
            Build_doctests

So people are likely to follow the warning and put in Build_doctests into their
Cabal file, and then be annoyed that it breaks their tests.

Because `Build_doctests` is autogenerated by Cabal-doctests, we know it will
never have doctests itself, so it can be safely filtered out.