Include non-absolute paths
The main motivation of the commits for issue #57 is to make it easier to use
Cabal-doctests with Haskell.nix.  Haskell.nix has built everything by the time
everything is run, so there are no intermediate directories around like `build/`
or `dist/`.

What Haskell.nix does instead is just make the original source available.  So it
seems convenient to include not just absolute paths, but relative paths as well.

Otherwise, we have to munge the "-i" flags after the fact.